feat: UI style changes and refactoring
