+++ src/CommandLine.cpp
@@ -1,6 +1,6 @@
 #include "CommandLine.h"
-#include "CommandRegister.h"
 #include <imgui.h>
+#include <util/CommandRegister.h>
 
 namespace ImGui
 {
@@ -16,28 +16,91 @@ namespace ImGui
 	void
 	CommandLine::DrawControlWindow() noexcept
 	{
-		if (ImGui::InputText("Command", &buf, ImGuiInputTextFlags_EnterReturnsTrue))
+		ImGui::Begin("Terminal");
+		ImGui::BeginChild(
+			"HistoryRegion",
+			ImVec2(0, -ImGui::GetFrameHeightWithSpacing()),
+			false,
+			ImGuiWindowFlags_HorizontalScrollbar);
+
+		for (const auto& [line, isError] : history)
+		{
+			if (isError)
+				ImGui::TextColored(ImVec4(1.0f, 0.2f, 0.2f, 1.0f), "%s", line.c_str());
+			else
+				ImGui::TextWrapped("%s", line.c_str());
+		}
+
+		if (ImGui::GetScrollY() >= ImGui::GetScrollMaxY())
+			ImGui::SetScrollHereY(1.0f);
+		ImGui::EndChild();
+
+
+		ImGui::Text(">>");
+		ImGui::SameLine();
+		ImGui::PushItemWidth(-FLT_MIN);
+
+		ImGuiInputTextFlags flags =
+			ImGuiInputTextFlags_EnterReturnsTrue | ImGuiInputTextFlags_CallbackHistory;
+
+		auto callback = [](ImGuiInputTextCallbackData* data) -> int {
+			CommandLine* console = reinterpret_cast<CommandLine*>(data->UserData);
+
+			if (data->EventFlag == ImGuiInputTextFlags_CallbackHistory)
+			{
+				if (data->EventKey == ImGuiKey_UpArrow)
+				{
+					if (console->historyIdx == -1)
+						console->historyIdx = static_cast<int>(console->history.size()) - 1;
+					else if (console->historyIdx > 0)
+						console->historyIdx--;
+
+					if (console->historyIdx >= 0 &&
+					    console->historyIdx < (int)console->history.size())
+						data->DeleteChars(0, data->BufTextLen),
+							data->InsertChars(
+								0,
+								console->history[console->historyIdx].first.c_str());
+				}
+				else if (data->EventKey == ImGuiKey_DownArrow)
+				{
+					if (console->historyIdx != -1 &&
+					    console->historyIdx < (int)console->history.size() - 1)
+						console->historyIdx++;
+					else
+						console->historyIdx = -1;
+
+					data->DeleteChars(0, data->BufTextLen);
+					if (console->historyIdx >= 0)
+						data->InsertChars(0, console->history[console->historyIdx].first.c_str());
+				}
+			}
+			return 0;
+		};
+
+		if (InputText("##command", &buf, flags, callback, this))
 		{
 			try
 			{
-				CommandRegister::Instance().Execute(buf);
-				history.push_back(buf);
+				util::CommandRegister::Execute(buf);
+				history.emplace_back(buf, false);
 			}
 			catch (const std::exception& e)
 			{
-				history.push_back(e.what());
+				history.emplace_back(e.what(), true);
 			}
 			catch (...)
 			{
-				history.push_back("Unknown exception");
+				history.emplace_back("Unknown exception", true);
 			}
+
 			buf.clear();
+			historyIdx = -1;
+
+			ImGui::SetKeyboardFocusHere(-1);
 		}
 
-		for (const auto& line : history)
-		{
-			ImGui::TextWrapped(line.c_str());
-		}
+		ImGui::End();
 	}
 
 	void
@@ -45,15 +108,15 @@ namespace ImGui
 	{
 		try
 		{
-			CommandRegister::Instance().Process();
+			util::CommandRegister::Process();
 		}
 		catch (const std::exception& e)
 		{
-			history.push_back(std::string("Command threw: ") + e.what());
+			history.emplace_back(std::string("Command threw: ") + e.what(), true);
 		}
 		catch (...)
 		{
-			history.push_back("Command threw unknown exception");
+			history.emplace_back("Command threw unknown exception", true);
 		}
 	}
 }
